test: extract the shared random-edit harness for the memo-freshness suites
Both memo-freshness differentials duplicated the PRNG, the op generator
and the transact/log/replay driver byte for byte, so a fix to the op
vocabulary had to be hand-synced and the suites could drift out of the
same edit distribution. The machinery now lives in
random-edit-test-support.ts; each suite keeps its own fixture, oracle
and op distribution.
