+++ packages/core/src/store/__tests__/note-scan-memo-freshness.test.ts
@@ -17,44 +17,20 @@
 //    plain descent (pre-truncation prefixes feed diagnoseNoteReferences).
 // 3. The budget-free memoized path, so the two memo systems cannot drift apart.
 //
-// Determinism: fixed seeds and a pure inline PRNG (mulberry32). A failure prints the
-// seed and the accumulated op script, which replays the exact sequence.
+// Determinism and the random-edit machinery live in `random-edit-test-support.ts`,
+// shared with `derivation-memo-freshness.test.ts`.
 
 import { describe, expect, test } from 'bun:test';
-import {
-  paragraphTextOf,
-  TreeDocumentStore,
-  type OoxmlElement,
-  type OoxmlNode,
-  type OoxmlPart,
-  type TreeDocOp,
-} from '../index.ts';
-import {
-  collectNoteReferences,
-  createNoteReferenceScanBudget,
-  readOoxmlPart,
-} from '../package/index.ts';
+import { TreeDocumentStore, type OoxmlPart } from '../index.ts';
+import { collectNoteReferences, createNoteReferenceScanBudget } from '../package/index.ts';
 import { budgetedNoteScanMemoStats } from '../package/note-references.ts';
-
-const W = 'http://schemas.openxmlformats.org/wordprocessingml/2006/main';
-
-function loadPart(xml: string): OoxmlPart {
-  const result = readOoxmlPart(xml, { name: '/word/document.xml', contentType: 'app/xml' });
-  if (!result.ok) throw new Error(result.reason);
-  return result.part;
-}
-
-/** Deterministic PRNG. Same seed, same sequence, every run, every machine. */
-function mulberry32(seed: number): () => number {
-  let state = seed >>> 0;
-  return () => {
-    state = (state + 0x6d2b79f5) >>> 0;
-    let t = state;
-    t = Math.imul(t ^ (t >>> 15), t | 1);
-    t ^= t + Math.imul(t ^ (t >>> 7), t | 61);
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
-}
+import {
+  driveRandomEdits,
+  loadPart,
+  mulberry32,
+  W,
+  type RandomOpKind,
+} from './random-edit-test-support.ts';
 
 // ── Fixture ────────────────────────────────────────────────────────────────────
 // Note references sit in every third paragraph, inside hyperlinks in some, so edits hit
@@ -76,82 +52,7 @@ function fixturePart(): OoxmlPart {
   );
 }
 
-// ── Reading the retained tree ──────────────────────────────────────────────────
-
-function* walk(node: OoxmlNode): Generator<OoxmlNode> {
-  yield node;
-  if (node.kind === 'textValue') return;
-  for (const child of node.children ?? []) yield* walk(child);
-}
-
-function paragraphSitesOf(part: OoxmlPart): { id: string; length: number }[] {
-  const sites: { id: string; length: number }[] = [];
-  for (const node of walk(part.root)) {
-    if (node.kind !== 'paragraph') continue;
-    sites.push({ id: node.id, length: paragraphTextOf(part, node.id)?.length ?? 0 });
-  }
-  return sites;
-}
-
-/** Adjacent (first, second) paragraph pairs among the body's direct children. */
-function adjacentBodyParagraphPairs(part: OoxmlPart): [string, string][] {
-  const body = part.root.children.find(
-    (child): child is OoxmlElement => child.kind !== 'textValue' && child.localName === 'body'
-  );
-  const pairs: [string, string][] = [];
-  const children = body?.children ?? [];
-  for (let index = 0; index + 1 < children.length; index += 1) {
-    const first = children[index]!;
-    const second = children[index + 1]!;
-    if (first.kind === 'paragraph' && second.kind === 'paragraph') {
-      pairs.push([first.id, second.id]);
-    }
-  }
-  return pairs;
-}
-
-// ── Op vocabulary ──────────────────────────────────────────────────────────────
-// Generated against the CURRENT tree, so ids and offsets are usually valid; the store
-// may still refuse. A refusal is logged and skipped, and the applied-op floor below
-// keeps the test from passing by refusing everything.
-
-type OpKind = 'insert-text' | 'delete-text' | 'split' | 'join';
-
-const OPS: readonly OpKind[] = ['insert-text', 'insert-text', 'delete-text', 'split', 'join'];
-
-function randomOp(part: OoxmlPart, rand: () => number, step: number): TreeDocOp | null {
-  const pick = <T>(items: readonly T[]): T => items[Math.floor(rand() * items.length)]!;
-  const sites = paragraphSitesOf(part);
-  if (sites.length === 0) return null;
-  const site = pick(sites);
-  switch (pick(OPS)) {
-    case 'insert-text':
-      return {
-        op: 'insertText',
-        paragraphId: site.id,
-        offset: Math.floor(rand() * (site.length + 1)),
-        text: ` step${step} extra`,
-      };
-    case 'delete-text': {
-      if (site.length < 2) return null;
-      const start = Math.floor(rand() * (site.length - 1));
-      const end = start + 1 + Math.floor(rand() * (site.length - start - 1));
-      return { op: 'deleteText', paragraphId: site.id, start, end };
-    }
-    case 'split':
-      return {
-        op: 'splitParagraph',
-        paragraphId: site.id,
-        offset: Math.floor(rand() * (site.length + 1)),
-      };
-    case 'join': {
-      const pairs = adjacentBodyParagraphPairs(part);
-      if (pairs.length === 0) return null;
-      const [firstId, secondId] = pick(pairs);
-      return { op: 'joinParagraphs', firstId, secondId };
-    }
-  }
-}
+const OPS: readonly RandomOpKind[] = ['insert-text', 'insert-text', 'delete-text', 'split', 'join'];
 
 // ── The oracle ─────────────────────────────────────────────────────────────────
 
@@ -254,36 +155,20 @@ const STEPS = 30;
 describe('budgeted note-reference scan memos stay fresh across random op sequences', () => {
   for (const seed of SEEDS) {
     test(`seed ${seed}: ${STEPS} random ops`, () => {
-      const rand = mulberry32(seed);
       const store = new TreeDocumentStore(fixturePart());
-      const log: unknown[] = [];
 
       // The fixture's typed references must have parsed, or the oracle compares empty
       // hit lists at every step.
       expect(collectNoteReferences(store.part).length).toBeGreaterThanOrEqual(6);
-      assertBudgetedScanFresh(store.part, seed, -1, log);
+      assertBudgetedScanFresh(store.part, seed, -1, []);
 
-      let applied = 0;
-      let splitsApplied = 0;
       const reusesBefore = budgetedNoteScanMemoStats.reuses;
-      for (let step = 0; step < STEPS; step += 1) {
-        const op = randomOp(store.part, rand, step);
-        if (op === null) {
-          log.push(['skip', step]);
-          continue;
-        }
-        const result = store.transact((tx) => {
-          tx.apply(op);
-        });
-        if (!result.ok) {
-          log.push(['refused', step, result.reason, op]);
-          continue;
-        }
-        log.push([step, op]);
-        applied += 1;
-        if (op.op === 'splitParagraph') splitsApplied += 1;
-        assertBudgetedScanFresh(store.part, seed, step, log);
-      }
+      const { applied, splitsApplied } = driveRandomEdits(store, {
+        rand: mulberry32(seed),
+        steps: STEPS,
+        ops: OPS,
+        onApplied: (step, _op, log) => assertBudgetedScanFresh(store.part, seed, step, log),
+      });
 
       // Anti-vacuity: enough ops applied, and enough splits that fresh spines were
       // rescanned against — and read back through — the memoized subtrees. The reuse
